Split event_handler.rs and menu_builder.rs into directory modules
- Convert event_handler.rs to event_handler/mod.rs + event_handler/tests.rs
- Convert menu_builder.rs to menu_builder/mod.rs + menu_builder/tests.rs
- All production code files now well under 500 LOC threshold
- Test modules extracted to separate files for better file health
